file is uploaded for the center of the polygon
A. Two Regular Polygons
time limit per test1 second
memory limit per test256 megabytes
inputstandard input
outputstandard output
You are given two integers n and m (m<n). Consider a convex regular polygon of n vertices. Recall that a regular polygon is a polygon that is equiangular (all angles are equal in measure) and equilateral (all sides have the same length).

Examples of convex regular polygons
Your task is to say if it is possible to build another convex regular polygon with m vertices such that its center coincides with the center of the initial polygon and each of its vertices is some vertex of the initial polygon.

You have to answer t independent test cases.

Input
The first line of the input contains one integer t (1≤t≤104) — the number of test cases.

The next t lines describe test cases. Each test case is given as two space-separated integers n and m (3≤m<n≤100) — the number of vertices in the initial polygon and the number of vertices in the polygon you want to build.

Output
For each test case, print the answer — "YES" (without quotes), if it is possible to build another convex regular polygon with m vertices such that its center coincides with the center of the initial polygon and each of its vertices is some vertex of the initial polygon and "NO" otherwise.

Example
inputCopy
2
6 3
7 3
outputCopy
YES
NO
Note
The first test case of the example
It can be shown that the answer for the second test case of the example is "NO".